% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checkGCDversion}
\alias{checkGCDversion}
\title{Check GCD package install}
\usage{
checkGCDversion()
}
\description{
Check if GCD package is installed and up to date to ensure always using the
most up to date GCD version. devtools package is required: on Windows
install Rtools.exe depending on your R version
http://cran.r-project.org/bin/windows/Rtools/
}
\details{
Last GCD database version is donwloaded and installed using:

\code{library(devtools)}

\code{install_github("GCD",username="paleofire",ref="master")}
}
\examples{

\dontrun{checkGCDversion()}

}
\author{
O. Blarquez
}
\keyword{~kwd1}
\keyword{~kwd2}

