\name{coast}
\alias{coast}
\docType{data}
\title{
coast}
\description{
World coastlines}
\usage{data(coast)}
\format{
  A data frame with 9865 observations on the following 2 variables.
  \describe{
    \item{\code{Y}}{Latitude}
    \item{\code{X}}{Longitude}
  }
}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}}
\examples{
data(coast)
}
\keyword{datasets}
