\name{paleofiredata}
\alias{paleofiredata}
\docType{data}
\title{
paleofiredata}
\description{
Charcoal data from the GCD v3}
\usage{data(paleofiredata)}
\format{
  A data frame with 118832 observations on the following 4 variables.
  \describe{
    \item{\code{ID_SITE}}{Sites Identifiers}
    \item{\code{DEPTH}}{Depths}
    \item{\code{EST_AGE}}{Estimated Age (cal yr BP)}
    \item{\code{QUANTITY}}{Charcoal Quantity}
  }
}

\source{
The Global Charcoal Databade v3}

\examples{
data(paleofiredata)
}
\keyword{datasets}
