\name{pfInteractive}
\alias{pfInteractive}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GCD sites interactive selection}
\description{
Interactive selection of GCD sites by drawing a polygon on a map.
}
\usage{
pfInteractive(addata=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{addata}{
An optional XY matrix of coordinates to specify a polygon to be drawn on the map.
}
}

\value{
An object of the class "pfSiteSel".
}
\author{
O. Blarquez}
\seealso{ \code{\link{pfSiteSel}}}

\examples{
## Type 
#ID=pfInteractive()
## And follow text instructions

}
