% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfPublication.R
\name{pfPublication}
\alias{pfPublication}
\title{Get citations for charcoal sites}
\usage{
pfPublication(x, output = "data.frame")
}
\arguments{
\item{x}{A "pfSiteSel" object}

\item{output}{Defines the output as a "list" or a "data.frame" (default).}
}
\value{
A list or data frame with citation informations related to charcoal sites.
}
\description{
Get citations for charcoal sites
}
\examples{
x=pfSiteSel(id_site \%in\% c(1:4))
pfPublication(x,output="list")
}
\author{
O. Blarquez
}
