% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potveg.R
\name{plot.potveg}
\alias{plot.potveg}
\title{plot.potveg}
\usage{
\method{plot}{potveg}(
  x,
  size = 4,
  palette = NULL,
  alpha = 0.5,
  text = FALSE,
  points = TRUE,
  ...
)
}
\arguments{
\item{x}{A "potveg object."}

\item{size}{Size of the dots on the map.}

\item{palette}{A custom color palette can be specified.}

\item{alpha}{Transparency of charcoal sites dots}

\item{text}{Logical: plot sites as numbers referring to potential vegetation
index (text=TRUE) or as points (text=FALSE, default).}

\item{points}{Logical: plot sites (TRUE, default)}

\item{\dots}{...}
}
\value{
A ggplot2 ("gg") object that can be further modified (see example)
}
\description{
Plot "potveg" object i.e. produce a map by overlaying charcoal sites on
potential vegetation maps. Uses ggplot2 syntax.
}
\examples{

\dontrun{
require(paleofire)
ID=pfSiteSel(c(1:10))
obj=potveg(ID,classif="l12")
plot(obj)

#Return a ggplot object
require(ggplot2)
p=plot(obj,text=TRUE,alpha=1)
p+ggtitle("My title")
}

}
\seealso{
\code{\link{potveg}}
}
\author{
O. Blarquez
}
