% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paleofire-internal}
\alias{bestLoess}
\alias{bootCI}
\alias{countries}
\alias{dem}
\alias{haverdist}
\alias{loessGCV}
\alias{paleofire-internal}
\title{Internal paleofire functions}
\description{
Internal paleofire functions and functions waiting for man.
}

