% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleofire-package.R
\name{paleofire-internal}
\alias{bestLoess}
\alias{bootCI}
\alias{countries}
\alias{dem}
\alias{haverdist}
\alias{influx}
\alias{loessGCV}
\alias{paleofire-internal}
\title{Internal paleofire functions}
\description{
Internal paleofire functions and functions waiting for man.
}

