% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pfInteractive}
\alias{pfInteractive}
\title{GCD sites interactive selection}
\usage{
pfInteractive(addata = NULL)
}
\arguments{
\item{addata}{An optional XY matrix of coordinates to specify a polygon to
be drawn on the map.}
}
\value{
An object of the class "pfSiteSel".
}
\description{
Interactive selection of GCD sites by drawing a polygon on a map.
}
\examples{
\dontrun{
# Type:
ID=pfInteractive()
# And follow text instructions}
}
\author{
O. Blarquez
}
\seealso{
\code{\link{pfSiteSel}}
}

