% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.contiguous}
\alias{plot.contiguous}
\title{Plot "contiguous" object}
\usage{
\method{plot}{contiguous}(x, ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{x}{An object returned by contiguous}

\item{ylim}{Numeric, ylim for the graph}

\item{xlim}{Numeric, xlim for the graph}

\item{\dots}{\dots{}}
}
\value{
A plot.
}
\description{
Plot an object returned by contiguous, plot contiguous cores (or sites) in
green (T) and non-contiguous cores in red (F).
}
\examples{
x=pfSiteSel(lat>12,lat<60,long<(-50),long>-140)
cont=contiguous(x)
plot(cont)
}
\author{
O. Blarquez
}
\seealso{
\code{\link{contiguous}}
}

