% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdffreq.R
\name{plot.kdffreq}
\alias{plot.kdffreq}
\title{plot.kdffreq}
\usage{
\method{plot}{kdffreq}(x, ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{x}{Object returned by kdffreq}

\item{ylim}{Numeric, y axis limits}

\item{xlim}{Numeric x axis limits}

\item{...}{other arguments}
}
\description{
Plot fire frequency calculated using the \code{\link[paleofire]{kdffreq}} function
}
\seealso{
\code{\link[paleofire]{kdffreq}} 
 @examples
 
 set.seed(123)
 fevent=c(round(abs(rnorm(20,mean=7,sd=5))*1000),round(abs(rnorm(10,mean=8,sd=1))*1000))

 ff=kdffreq(fevent,bandwidth = 1000, nbboot=10)
 plot(ff)
}

