% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfSiteSel.R
\name{plot.pfSiteSel}
\alias{plot.pfSiteSel}
\title{plot.pfSiteSel}
\usage{
\method{plot}{pfSiteSel}(x, add = NULL, type = "Map", zoom = "Sites",
  pch = "|", xlim = NULL, ylim = NULL, cex = 1,
  plot_countries = FALSE, ...)
}
\arguments{
\item{x}{An object of the class "pfSiteSel".}

\item{add}{An object returned by pfAddData (optional).}

\item{type}{Character, type of plot among "Map" or "Chronology".}

\item{zoom}{Character, zooming factor for type="Map": "Sites" or "World"}

\item{pch}{Pointer type see \code{\link[graphics]{plot}}.}

\item{xlim}{Numeric, x axis limits.}

\item{ylim}{Numeric, y axis limits.}

\item{cex}{Numeric, size of points.}

\item{plot_countries}{Logical, default FALSE (if TRUE plot countries borderlines and coastlines)}

\item{\dots}{\dots{}}
}
\description{
Plot an object of the class "pfSiteSel"
}
\examples{

ID=pfSiteSel(id_region=="ENA0", long>-100)
plot(ID,zoom="world")


}
\author{
O. Blarquez
}

