% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pfSiteSel.R
\name{summary.pfSiteSel}
\alias{summary.pfSiteSel}
\title{summary.pfSiteSel}
\usage{
\method{summary}{pfSiteSel}(object, ...)
}
\arguments{
\item{object}{An object of the class "pfSiteSel".}

\item{\dots}{\dots{}}
}
\value{
Data.frame, returns the following informations: "id_site", "lat",
"long" "elev", "min_est_age", "max_est_age", "num_dating", "date_int",
"num_samp", "l12", "rf99".
}
\description{
Return a summary table for an object of the class "pfSiteSel"
}
\examples{
ID=pfSiteSel(id_site==2)
summary(ID)
}
\author{
O. Blarquez
}

